BareMetal: Allow to choose 'tools.ini' for UVSC provider configuration
An each Keil installation contains an own 'tools.ini' file.
The "tools.ini" file is used to determine the paths to the installed
debugger drivers as well as to the device software packs. These drivers
and packs are displayed on a UVSC provide configuration widget and are
used for debugging.

Previously, a path to 'tools.ini' file was detected automatically
from the configured Kit of the opened project, using the
SessionManager::startupTarget() method (a Kit contains the debugger
path, which is the uVision executable).

But, in that case the user can't configure the UVSC provider if a
project is not opened yet.

This issue is solved by adding a path chooser to the UVSC provider
configuration widget. That allows to user to choose the desired
'tools.ini' file instance in any time.

Change-Id: Ife44d69704c9dd2102baecb0ff1954fa4bc05f1e
Reviewed-by: hjk <hjk@qt.io>