cmake build: Fix qdoc call for LLVM-based qdoc
It requires us to pass a whole bunch of include paths manually,
including Qt include paths.

Extracts a stripped down list of include paths from all the plugin and
library targets that we know of.
On macOS, also gets Qt's framework path derived from the QtCore location.

Since these can contain generator expressions, we have to write them to
a file (so the expressions are resolved).
We pass this file with qdoc options with the hidden "@" command line
feature of qdoc.

Task-number: QTCREATORBUG-22451
Change-Id: Ifae6960023cc6e63cd66104417dd4a16f2e491a2
Reviewed-by: Cristian Adam <cristian.adam@qt.io>