MsvcToolchain: Rework detection of supportedAbis
Commit 9912a409da107848b2876d75649c250b37af7c80 made changes to
implement user MSVC toolchains allowing to set custom arguments to
vcvars bat file.

This commit introduced the possibility to have multiple supported Abis
for a MSVC toolchain.
But the supported ABI detection broke with older toolchain that are
initialized with a vcvars32.bat and the like (as defined in
MsvcPlatform platforms).

msvc2010 and VC2015 build tools got affected, the detected Abis were
empty and made Qt Creator to fail to load these toolchain from
toolchain.xml (as a non empty m_supportedAbis list was required).

The effect of this is that any Kit that was referring to these compilers
was reset to "No compiler" every time Qt Creator is started.
In that case, the user could workaround the issue by reconfiguring Kits
after each Qt Creator startup or never close QtCreator which is
impractical.

This commit is a rework of the fix introduced in
708179285de9f58e8578b72f7320c37b6252c45f which was a minimal fix
introducted in 4.10.2.

This commit fix this issue by removing the notion of multiple ABI per
MSVC toolchain and instead iterate all detected toolchain for matching
ones when populating the ABI widget via setAbis.

This also makes only one proposed ABI in the mainComboBox of AbiWidget
for any selected varsBatArchCombo. So it is no more possible to have a
32 bits proposed ABI with a /x64 varsBatArchCombo for example.

At any time, if the user wants to set a different ABI for a given
varsBatPath/varsBatArch, he can still select the "custom" item and
manually configure each part of the ABI.

Task-number: QTCREATORBUG-22960
Change-Id: I753e96db3b0e83371fa39d91c4a3592bd1552b8e
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>