TextEditor: Fix scrolling state after splitting
And when restoring editors in other ways (like from a session).

When splitting, the new editor was always scrolling to make the text
cursor visible, even if it wasn't visible in the original.

1. open editor with long contents
2. scroll away from the text cursor, so the line isn't visible
3. split and wonder

or 3. close and re-open Qt Creator, and restore the session

Our code already does set the state of the new editor to the old one,
but that state was destroyed by QPlainTextEdit::showEvent, which scrolls
on first show if text was added to the document.
So, wrap the showEvent with save/restoreState on first show.

Change-Id: I95e0b4e963ebb33d13ce020affaf7fc7f94c9257
Reviewed-by: David Schulz <david.schulz@qt.io>