Qmake: Use buildableFileProFile(node) for Build File action
Should be at least theoretically the right one, but is mainly
done to always feed a QmakeProFileNode * to the 5-args
handleSubDirContextMenu().

In practice it does not seem to impact the current level of brokenness:
In case of Build->File with not-synchronized editors the action shows
the name of the file selected in the project view, while it actually
builds the file in the current editor.

Change-Id: I47f8f22eb80be18d1be5997a861cea89bd069d1b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>