ProjectExplorer: Trigger updates more directly
Changes of active run and build config impact some visible state, but
instead of connecting and disconnecting signals to the target, make the
call explicit.

Also, triggering only updateAction is sufficient, as this will already
run updateRunAction. This makes the number of doUpdateRunAction calls
when opening a normal QMake project with one target, three build configs
drops from 17 to 15.

There was some mechanism on the receiver side originally meant to
prevent excessive updates, which broke with 7ae3589 when remembering the
previous configurations was accidentally removed.

This here drops the mechanism completely, as active changed signals
would not trigger twice in succession for the same config anyway.

Change-Id: Id9f87e542d8871733966ab79118734b7e9ad1e8a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>