QmlDesigner: Use ActionManager in FormEditor
We assign an ICore context to the FormEditorWidget and
associate it with a Context (C_QMLFORMEDITOR).
This guarantees that the action are only active when
the FormEditorWidget has focus.
The actions are then registered as Commands. The shortcuts are set on the
command and are configurable in the options of Qt Creator.

Task-number: QDS-1712
Change-Id: I327f35736205c9337386cf70e6f058d4b19be50e
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>