McuSupport: If kits are missing, provide an InfoBar that opens settings
When installing Qt for MCUs via the SDK installer, the user (in many
cases first-time evaluator) has to create the kits for Qt for MCUs. For
those who see Qt Creator for the first time, it is difficult to find out
where to search for the UI that generates those kits.

Let's help the user with an InfoBar entry that in case of ...

 * McuSupport plugin being enabled (by SDK installer)
 * Settings key [McuSupport]/Package_QtForMCUsSdk set (by SDK installer)
 * Absence of valid Qt for MCUs Kits

... helps to directly open the right page and tab of the Options dialog.

Task-number: QTCREATORBUG-23943
Change-Id: I7d4834633b1b94ef663d62cd395ee071bd123f1f
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>