BareMetal: Add support for peripheral registers view using UVSC provider
Now it is possible to use the SVD files to read or write the peripheral
registers of a target device.

By default a paths to the SVD files automatically comes from the installed
"Device Software Pack" database and don't need to be changed by the user.
But, if a user needs to change the SVD file, then it can be done via the
provider's device selection options.

Tested on Windows with Keil MDK-ARM v5.23 using the STM32 NUCLEO-F767ZI
board.

Change-Id: Ia39e8d9a25cd24461804d47830ee9e8f01486108
Reviewed-by: hjk <hjk@qt.io>