Fix split side by side behavior again
After fixing top/bottom splits, the side by side splitting scrolled
wrongly if the cursor position was somewhere at the bottom of the view.

Avoid restoring state if the editor is not in a layout, since the
behavior can now differ depending on it. Calling restoreState before
that, would make the behavior depend on the default size of the widget.

This has the additional benefit that editors can adapt (e.g. make a
previously visible text cursor visible) if they are moved/duplicated
when opening a document in an existing split, not only when creating the
split.

Change-Id: I3916eb20c2251dbeac4b604f96af647cf142fcad
Reviewed-by: David Schulz <david.schulz@qt.io>