BareMetal: Fix parsing of software pack description files for UVSC provider
A target device info locates in the *.pdsc files provided by the device
description packages which are installed into the Keil MDK-ARM instance.

We parse this files to provide possibility for the user to choose a
specific MCU target for debugging.

Previous parsing implementation was complicated and did not take in
account the specific *.pdsc format features:

 * https://www.keil.com/pack/doc/CMSIS/Pack/html/packFormat.html

It turns out that the following entries as 'family', 'sub-family',
'device', and 'device-variant' may contain an attributes with the
same names. So, we can simplify creation of an items, use only one
base implementation for all items. Besides, a new implementation
simplifies parsing and for other *.pdsc files (e.g. non only for
STMicroelectronics).

Tested with Keil MDK-ARM v5.23 using STM32 NUCLEO-F767ZI board.

Change-Id: Idd51b8fa3e36a7896254b4cb8aa4c1287b9f0df5
Reviewed-by: hjk <hjk@qt.io>