QbsPM: Fix wrong {GCC|Clang} compiler name and prefix handling
When the GCC or Clang compiler name ends with the version sub-string
(f.e. arm-none-eabi-gcc-8.2.1.exe), this leads to the wrong cCompilerName
detection, that causes the project parsing errors and disables the project:

  The following properties are not set. Set them in your profile or product:
  cpp.compilerIncludePaths
  cpp.compilerFrameworkPaths
  cpp.compilerLibraryPaths

  Product 'xyz' had errors and was disabled.

Now we return a proper prefix (e.g. arm-none-eabi-) and cCompilerName
with a version (e.g. 'gcc-8.2.1.exe')

Change-Id: I5dd42a343a0982326ed0f23b821e5016b8df39f1
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>