QmlDesigner: Fix id corruption on paste
When pasting model nodes, any bound expression referring to a node in
the same pasted set gets fixed so that the expression refers to a
pasted copy of the node instead.
This fix makes it so that only fully matching ids are replaced instead
of also partial id matches. The old matching broke e.g. when the paste
included both "cube" and "cubeMaterial" and a binding to
"cubeMaterial". Without enforcing the full id matching, the binding
became "cube1Material1".

Change-Id: I3c622ea946c74d59a192157ca06d096467428364
Fixes: QDS-1564
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>