QmlDesigner: introduce DisableItemLibraryUpdateTimer
This is a temporary change to enable better tracing of a bug
which we have for a long time.
Be aware if this is set it can be slow down importing or open
bigger projects/files in the QtQuick DesignMode.

add this under the category [QML] to your settings file:
Designer\DisableItemLibraryUpdateTimer=true

Change-Id: I0919fe18e358e68905eb4a7f4849c6841f26a10d
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>