QmlDesigner: Enable setting a material's diffuseMap by drag and drop
Dragging and dropping an image from the assets on a default material
creates a texture and sets its source the the image and the applies
the texture to the diffuseMap property of the material making it
pretty straight forward to set a texture on a marerial.

Additionally dropping an image on a texture updates the texture's source
property instead of creating an image object.

Task-number: QDS-1844
Change-Id: I6906851ad070e06a17b1ef7d4321e302bb305d29
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>