QmlDesigner: Fix ItemLibraryEntry serialization
When data is read from QDataStream into Qt containers, it is simply
appended/inserted into the container rather than completely overwriting
the container contents. This isn't desirable when reading an
ItemLibraryEntry object from data stream, as the assumption is that we
want the entry to be exactly what was originally stored into the
stream.

Task-number: QDS-1815
Change-Id: Ifaee244b2b4e7326b6f10fde3faeb5218f505ee2
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>