Support adding multiple shortcuts to the same action
In the settings UI. Moves the "Reset" button for individual
actions up to under the list, next to the "Reset All" button.

Users can add a new "row" for another shortcut for the same action,
as long as there are no shortcut inputs empty.
There is no way to directly remove an input row - to remove a shortcut,
just clear the input (like it was the case with just single shortcuts).
This gets cleaned up when you select an item again.

Fixes: QTCREATORBUG-72
Change-Id: Id0402d00ebeb41f5b0c612d9d03f884b78485fbc
Reviewed-by: David Schulz <david.schulz@qt.io>