AutoTest: Pull catch tree nodes one level up
This patch makes test cases now match the TestCase node so that possible
additional sections can be functions instead of test data.
Mis-use the TestSuite for the file node that is used to additionally
group the items. The file node is kept - for now - as it seems rather
often that test case names get pretty long names.

Change-Id: I364fc40d1de6a22b6f170b3001f37393b64680aa
Reviewed-by: David Schulz <david.schulz@qt.io>