Theme: add design studio controls theming to theme color list
I've chosen to add the controls themes colors to the main color list for couple of reasons.

1.) Adding the colors to the main color list allows them to be used from the QML constants
file that defines the controls color theme, in turn allowing us to customize the look and
feel of all controls on a per theme basis.

2.) I have added new colors instead of using the existing ones because of the ripple effects
of modifying existing colors in the themes, i.e that often when you want to change one particular
color in the UI modifying the value of existing colors in the theme has the unintended consequence
of changing other parts of the UI. Thus adding a specific set for the controls allows us to have a
fine grained approach to modifying the controls in any theme without disturbing the other elements
of the existing UI theme.


Change-Id: I16fb0458c870d01f68ed626676a665117628c917
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>