EmacsKeys: Replace U+2029 with LF when copying text to clipboard
Otherwise '\n' would be represented as U+2029 in clipboard, which is
useless.

Some of the cursor.selectedText() calls could be replaced with
QPlainTextEdit::copy(), which is backed by createMimeDataFromSelection(),
but there's a slight behavior difference such as EmacsKeysPlugin::copy()
does clear the selection. So I think it's better to consistently use the
lower-layer function.

Change-Id: Ida06b0c250774d80fb729224fd9492a35ea59b7d
Reviewed-by: David Schulz <david.schulz@qt.io>