Debugger: Minimize stalling at break-points insertion in UVSC engine
Sometime the UVSC_DBG_CREATE_BP function stalls within 10 seconds
and then returns with the timeout error.

But this problem can be avoided by creating and then removing the
fake breakpoint before creating the original breakpoint. In this case,
we need to remove the line number from the expression of this fake
breakpoint.

Of course, it does not help when the breakpoint cannot be created
with the specified expression. In this case the UVSC_DBG_CREATE_BP
function will stall anyway within 10 seconds.

Change-Id: Ie8d39c545683d6bb3e29e5bb7c4ac0fd1e34222a
Reviewed-by: hjk <hjk@qt.io>